add RespoState derive macro as sugar
