+++ .gitattributes
@@ -1,7 +1,7 @@
 
 
 yarn.lock -diff linguist-generated
-Cargo.lock -diff linguist-generated
+# Cargo.lock -diff linguist-generated
 lib -diff linguist-generated
 
 examples/snippet_*.rs -diff linguist-generated
