+++ demo_respo/Cargo.toml
@@ -20,6 +20,7 @@ console_error_panic_hook = "0.1.7"
 uuid = { version = "1.8.0", features = [ "v4", "js" ] }
 serde = { version = "1.0.203", features = ["derive", "rc"] }
 serde_json = "1.0.117"
+respo_state_derive = { path = "../respo_state_derive" }
 
 [dependencies.web-sys]
 version = "0.3.69"
