+++ respo/Cargo.toml
@@ -20,6 +20,8 @@ cirru_parser = "0.1.31"
 rust-hsluv = "0.1.4"
 serde = { version = "1.0.203", features = ["derive", "rc"] }
 serde_json = "1.0.117"
+respo_state_derive = { path = "../respo_state_derive" }
+
 
 [lib]
 crate-type = ["cdylib", "rlib"]
