+++ respo/src/dialog/prompt.rs
@@ -5,6 +5,7 @@ use std::marker::PhantomData;
 use std::rc::Rc;
 
 use js_sys::{Array, Function, Reflect};
+use respo_state_derive::RespoState;
 use serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::Closure;
 use wasm_bindgen::{JsCast, JsValue};
@@ -67,22 +68,10 @@ impl PromptValidator {
   }
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize, RespoState)]
 struct InputState {
   draft: String,
   error: Option<String>,
-}
-
-impl RespoState for InputState {
-  fn backup(&self) -> Option<serde_json::Value> {
-    serde_json::to_value(self).ok()
-  }
-
-  fn restore_from(&mut self, s: &serde_json::Value) -> Result<(), String> {
-    let v: Self = serde_json::from_value(s.to_owned()).expect("failed to deserialize");
-    *self = v;
-    Ok(())
-  }
 }
 
 fn comp_prompt_modal<T, U, V>(
@@ -265,25 +254,10 @@ where
   fn share_with_ref(&self) -> Rc<Self>;
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize, RespoState)]
 struct PromptPluginState {
   show: bool,
   text: Option<String>,
-}
-
-impl RespoState for PromptPluginState {
-  fn backup(&self) -> Option<serde_json::Value> {
-    serde_json::to_value(self).ok()
-  }
-  fn restore_from(&mut self, s: &serde_json::Value) -> Result<(), String> {
-    match serde_json::from_value(s.to_owned()) {
-      Ok(v) => {
-        *self = v;
-        Ok(())
-      }
-      Err(e) => Err(format!("failed to deserialize: {:?}", e)),
-    }
-  }
 }
 
 /// a dialog for prompt, request for some input, and submit
