+++ respo_state_derive/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "respo_state_derive"
+version = "0.0.1"
+edition = "2021"
+
+[dependencies]
+quote = { version = "1.0.36", features = ["proc-macro"] }
+syn = { version = "1.0.109", features = ["clone-impls", "derive", "parsing", "printing", "proc-macro"] }
+
+# proc-macro2 = "1.0.85"
+# proc-macro2 = { version = "1.0.85", features = ["proc-macro"] }
+
+[lib]
+proc-macro = true
